api/voiture/requirements.txt
Django~=4.2
djangorestframework~=3.14.0
psycopg2-binary~=2.9.6
requests~=2.25.1
Pillow~=9.0.1
gunicorn==20.0.4